Extract logic to use-question-editor-sync.ts
